minecraft-controller/src/requirements.txt
flask
mctools
python-dotenv